Codes for Question 1:
Described instacart dataset.
Determined how many aisles and which items had the most ordered items.
Made a plot that shows the number of items ordered in each aisle, limiting this to aisles with more than 10,000.
Made a table showing the three most popular items in each of the aisles “baking ingredients”, “dog food care”, and “packaged vegetables fruits”.
Made a table showing the mean hour of the day at which Pink Lady Apples and Coffee Ice Cream are ordered on each day of the week.